pyproject.toml
[tool.poetry]
name = "grafana-query-tool"
version = "0.1.0"
description = ""
authors = ["showchan33 <showchan3333@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
playwright = "^1.41.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
